- Service layer implemented using Java 8 streams. - Tests updated for better coverage

+++ src/main/java/com/acme/addressbook/service/AgeService.java
@@ -4,5 +4,5 @@ import com.acme.addressbook.model.Person;
 
 public interface AgeService {
     Person getOldestPerson();
-    int getAgeDiffInDaysByName(String nameA, String nameB);
+    long getAgeDiffInDaysByName(String nameA, String nameB);
 }
